Backend/requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
SQLAlchemy==1.3.1
boto3==1.16.25
botocore==1.19.25
click==7.1.2
itsdangerous==0.24
Jinja2==2.10
jmespath==0.10.0
MarkupSafe==1.1.1
Pillow==8.0.1
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.26.2
Werkzeug==0.14.1